Updated USLM-2_0-Review-Guide.md for 2.0.4
Signature
The model for <signatures> was modified to better match actual signature instances (previously unused). Elements were added for <notation> and <autograph>. In addition, <signatures> are now allowed at the end of levels to support the Federal Register.
…
Amendments
In order to allow the use of the element <action> in the context of bill actions, the existing use of <action> within amendments has been changed to <amendingAction>.
The existing action "renumber" was renamed "redesignate" to better match the terminology used in Congress. Additional actions of "conform", "repealAndReserve" and "unknown" were added.

PDF of Review Guide has not been updated yet.